Overhauling component and style structure to make code for more code intensive pages more compoact and modularized.
