+++ legacy/components/confirmation_components/AttendSuccess.js
