+++ legacy/components/confirmation_components/CreateEventSuccess.js
