+++ legacy/components/display_components/AttendanceChart.js
