+++ legacy/styles/display_styling/attendance_chart.scss
