+++ legacy/styles/display_styling/event_details.scss
