+++ legacy/styles/input_styling/text_area.scss
