+++ legacy/styles/input_styling/time_drop_down.scss
