+++ legacy/styles/transition_styling/share_to_grids.scss
