+++ src/styles/accounts_styling/forgot_Password/forgotPassword.scss
