+++ src/styles/create_pages_styling/create_page/create.scss
@@ -1,7 +1,7 @@
 //Imports
-@use './../variables' as v;
-@use './../mixins' as m;
-@use './../device' as d;
+@use './../../variables' as v;
+@use './../../mixins' as m;
+@use './../../device' as d;
 
 //%&%&%&%&%&%&%&%&%&%&%&%&%&%&%&%&%&%&%&% Container Styles %&%&%&%&%&%&%&%&%&%&%&%&%&%&%&%&%&%&%&%
 #create-shell {
