+++ src/styles/home_styling/home.scss
